Исправлена ошибка:
System.NullReferenceException: Ссылка на объект не указывает на экземпляр объекта.
в Inforoom.PriceProcessor.Downloader.RostaCertificateCatalogHandler.ProcessData()

+++ src/PriceProcessor/Downloader/RostaCertificateCatalogHandler.cs
@@ -40,7 +40,7 @@ namespace Inforoom.PriceProcessor.Downloader
 							ImportCatalogFile(catalogFile);
 					}
 					finally {
-						if (File.Exists(catalogFile.LocalFileName))
+						if (catalogFile != null && File.Exists(catalogFile.LocalFileName))
 							File.Delete(catalogFile.LocalFileName);
 					}
 				}
